Move catchup logic in patch player to CatchupManager
