+++ firmware/src/patch_play/patch_player.hh
@@ -9,6 +9,7 @@
 #include "midi/midi_message.hh"
 #include "midi/midi_router.hh"
 #include "null_module.hh"
+#include "params/catchup_manager.hh"
 #include "params/catchup_param.hh"
 #include "patch/midi_def.hh"
 #include "patch/patch.hh"
@@ -67,15 +68,8 @@ private:
 	uint32_t midi_divclk_ctr = 0;
 	uint32_t midi_divclk_div_amt = 0;
 
-	// knob_maps[knobset_idx][panel_knob_id].map and .catchup
-	struct MappedParam {
-		MappedKnob map;
-		CatchupParam catchup;
-	};
-	using ParamSet = std::array<std::vector<MappedParam>, PanelDef::NumKnobs>;
 	std::array<ParamSet, MaxKnobSets> knob_maps;
-
-	std::array<float, PanelDef::NumKnobs> panel_knobs{};
+	CatchupManager catchup_manager;
 
 	std::array<bool, NumOutJacks> out_patched{};
 	std::array<bool, NumInJacks> in_patched{};
@@ -285,22 +279,21 @@ public:
 	// K-rate setters/getters:
 
 	void set_panel_param(unsigned panel_knob_id, float val) {
-		panel_knobs[panel_knob_id] = val;
-
-		for (auto &knob_map : knob_maps[active_knob_set][panel_knob_id]) {
-			auto &k = knob_map.map;
-			auto &catchup = knob_map.catchup;
-
-			auto module_val = modules[k.module_id]->get_param(k.param_id);
-			auto scaled_phys_val = k.get_mapped_val(val);
-			if (auto v = catchup.update(scaled_phys_val, module_val)) {
-				modules[k.module_id]->set_param(k.param_id, *v);
-			}
-		}
+		catchup_manager.set_panel_param(modules, knob_maps[active_knob_set], panel_knob_id, val);
 	}
 
 	void set_panel_input(unsigned jack_id, float val) {
 		set_all_connected_jacks(in_conns[jack_id], val);
+	}
+
+	void set_active_knob_set(unsigned num) {
+		auto new_active_knob_set = std::min(num, MaxKnobSets - 1);
+
+		if (active_knob_set != new_active_knob_set) {
+			active_knob_set = new_active_knob_set;
+
+			catchup_manager.reset(modules, knob_maps[active_knob_set]);
+		}
 	}
 
 	void send_raw_midi(MidiMessage msg) {
@@ -480,24 +473,6 @@ public:
 	void add_midi_mapped_knob(const MappedKnob &map) {
 		if (pd.add_update_midi_map(map)) {
 			cache_midi_mapping(map);
-		}
-	}
-
-	void set_active_knob_set(unsigned num) {
-		auto new_active_knob_set = std::min(num, MaxKnobSets - 1);
-
-		if (active_knob_set != new_active_knob_set) {
-			active_knob_set = new_active_knob_set;
-
-			// Reset all catchups in the new active knobset.
-			// This allows them to change to catchup mode if necessary
-			for (unsigned i = 0u; auto &knob : knob_maps[active_knob_set]) {
-				for (auto &map : knob) {
-					auto module_val = modules[map.map.module_id]->get_param(map.map.param_id);
-					map.catchup.reset_phys_val(panel_knobs[i], module_val);
-				}
-				i++;
-			}
 		}
 	}
